wsd: test: fold UnitWOPILoadEncoded test into UnitWOPI
The load-encoded is a very small and quick
test that only loads and immediately closes
a document with an special-character in the
filename.

There is no need to have a separate unitlib
for it and load a whole wsd etc.

Also, we fix a bug that resulted in multiple
documents being loaded, as we didn't transition
away from the Load status. This was a hidden
performance cost, but also introduced instability.
